Update input tensor on model training of different size
fixed the following issue:
If the inference was previously running a two-handed model and then a one hand model is trained/received, it will be feeding two-handed data into the one-handed model (or vice-versa), which for some reason still works (don’t know why), but probably causes some issues. In the UGInferenceRunnerScript, add functionality to update the inferenceHandMode and inputTensor after receiving a new model when necessary.